+++ index.php
@@ -29,7 +29,7 @@ require_once(__DIR__ . '/../../config.php');
 global $OUTPUT, $PAGE;
 
 $courseid = required_param('courseid', PARAM_INT);
-$excludedisabledcms = optional_param('edc', false, PARAM_BOOL);
+$excludedisabledcms = optional_param('edc', true, PARAM_BOOL);
 $ctx = context_course::instance($courseid);
 $course = get_course($courseid);
 
@@ -72,6 +72,7 @@ foreach (mod_util::get_cms_with_metadata($courseid, $excludedisabledcms) as $obj
         'ready' => $obj->ready,
         'canbearchived' => $obj->supported && $obj->enabled && $obj->ready,
         'lastarchived' => $obj->lastarchived ?: null,
+        'dirty' => $obj->dirty,
     ];
 }