+++ tests/activity_archiving_task_test.php
@@ -17,6 +17,7 @@
 namespace local_archiving;
 
 use local_archiving\type\activity_archiving_task_status;
+use local_archiving\type\cm_state_fingerprint;
 use local_archiving\type\db_table;
 use local_archiving\type\filearea;
 use local_archiving\type\task_content_metadata;
@@ -61,9 +62,11 @@ final class activity_archiving_task_test extends \advanced_testcase {
         $job = $this->generator()->create_archive_job();
 
         // Create a new activity archiving task.
+        $fingerprint = cm_state_fingerprint::from_raw_value(str_repeat('1234', 16));
         $task = activity_archiving_task::create(
             jobid: $job->get_id(),
             context: $job->get_context(),
+            cmfingerprint: $fingerprint,
             userid: $job->get_userid(),
             archivingmodname: 'quiz',
             settings: (object) ['foo' => 'bar'],
@@ -74,6 +77,7 @@ final class activity_archiving_task_test extends \advanced_testcase {
         $this->assertSame($job->get_userid(), $task->get_userid(), 'User ID should match');
         $this->assertSame('quiz', $task->get_archivingmodname(), 'Archiving mod name should match');
         $this->assertEquals($job, $task->get_job(), 'Job should match');
+        $this->assertEquals($fingerprint, $task->get_fingerprint(), 'CM state fingerprint should match');
 
         // Check that the task can be retrieved by its ID.
         $retrievedtask = activity_archiving_task::get_by_id($task->get_id());
@@ -600,6 +604,52 @@ final class activity_archiving_task_test extends \advanced_testcase {
         $this->assertNotNull($fileinfo->itemid, 'Item ID should exist');
         $this->assertNotEmpty($fileinfo->filepath, 'Filepath should not be empty');
         $this->assertSame('testfile.txt', $fileinfo->filename, 'Filename should match provided name');
+    }
+
+    /**
+     * Tests checking for existing fingerprints in activity archiving tasks.
+     *
+     * @covers \local_archiving\activity_archiving_task
+     *
+     * @return void
+     * @throws \coding_exception
+     * @throws \dml_exception
+     * @throws \moodle_exception
+     */
+    public function test_fingerprint_exists(): void {
+        // Prepare.
+        $this->resetAfterTest();
+        $course = $this->generator()->create_course();
+        $cm = $this->generator()->create_module('quiz', ['course' => $course->id]);
+        $ctx = \context_module::instance($cm->cmid);
+        $driver = \local_archiving\driver\factory::activity_archiving_driver('quiz', $ctx);
+
+        // There should be no fingerprint for a freshly created cm.
+        $this->assertFalse(
+            activity_archiving_task::fingerprint_exists($ctx, $driver->fingerprint()),
+            'No fingerprint should exist for a fresh cm without any activity archiving tasks'
+        );
+
+        // Uncompleted tasks should not be considered as existing fingerprints by default.
+        $task = $this->generator()->create_activity_archiving_task();
+        $this->assertFalse(
+            activity_archiving_task::fingerprint_exists($task->get_context(), $task->get_fingerprint()),
+            'Fingerprint should not be considered as existing for uncompleted tasks'
+        );
+
+        // We should be able to force the fingerprint check to ignore task status.
+        $this->assertTrue(activity_archiving_task::fingerprint_exists(
+            $task->get_context(),
+            $task->get_fingerprint(),
+            requiresuccess: false
+        ), 'Fingerprint should be considered as existing when ignoring task status');
+
+        // Once the task is completed, the fingerprint should be considered as existing.
+        $task->set_status(activity_archiving_task_status::FINISHED);
+        $this->assertTrue(
+            activity_archiving_task::fingerprint_exists($task->get_context(), $task->get_fingerprint()),
+            'Fingerprint should be considered as existing for completed tasks'
+        );
     }
 
 }